+++ RELEASE-NOTES.md
@@ -57,6 +57,8 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Python: Adds the `AXOM_PYTHON_MODULE_INSTALL_PREFIX` CMake variable to control where Axom installs its Python
   package(s), relative to the install prefix.
 - Quest: The C2CReader was enhanced to provide assembly support.
+- Core: Adds `utilities::filesystem::getFileExtension(str)`
+- Klee: Adds support for lua-based input decks for shaping
 
 ### Removed
 - Bump: Removed `axom::bump::views::MultiBufferMaterialView`, which was a view type for an obsolete flavor of Blueprint matset.